refactor: extract pagination helper and move writeJSON/writeError to helpers.go
- Created internal/api/routes/helpers.go with parseIntParam(), writeJSON(),
  and writeError()
- Replaced duplicated limit/offset parsing in episodes.go, memories.go,
  patterns.go, and llm_usage.go with parseIntParam() calls
- Moved writeJSON/writeError out of memories.go into the shared helpers file

Closes #89
Closes #92
